Fix for iOS 18 Safari losing window.parent
The broadcast channel is only sent data when window.parent isn't available.

One security implication - any frame on the same origin can create a brodcast channel and receive the messages. We are using the randomly generated hellos.js callbackId as the channel name to mitigate this. A longer string might be better.

https://kamihq.slack.com/archives/C7RP9R1QC/p1728503865166379